delete todo validation added

+++ todo-app/src/todos/app.js
@@ -47,7 +47,10 @@ export const App = (elementId) => {
     })
 
     todoListUL.addEventListener('click',(event) =>{
+        const idDestroyElement = event.target.className === 'destroy';
         const element = event.target.closest('[data-id]');
+
+        if( !element || !idDestroyElement ) return;
         todoStore.deleteTodo(element.getAttribute('data-id'));
         displayTodos();
     })
